feat(recipe/spin/): update config validation
